requirements.txt
openai>=1.0.0
chromadb>=0.4.0
PyYAML>=6.0

# If you see errors with chromadb Settings-based init, try upgrading chromadb:
# pip install --upgrade chromadb
